Update README with steps that lead to a working build
There are either some implicit steps missing from the README, or some things have changed in the repo over time that made the instructions not as useful as they could be. Add more verbosity to the README to help users get to a functional site that they can iterate on.